@@ -1,6 +1,7 @@
 package net.lucode.hackware.magicindicator;
 
-import android.support.v4.view.ViewPager;
+
+import androidx.viewpager.widget.ViewPager;
 
 /**
  * 简化和ViewPager绑定
