@@ -1,13 +1,13 @@
 package com.trs.channellib.channel.channel;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.helper.ItemTouchHelper;
 import android.util.AttributeSet;
 import android.widget.FrameLayout;
+
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.ItemTouchHelper;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.trs.channellib.R;
 import com.trs.channellib.channel.channel.helper.ItemDragHelperCallback;
