@@ -2,13 +2,14 @@ package com.trs.channellib.channel.channel.helper;
 
 import android.content.Context;
 import android.graphics.drawable.BitmapDrawable;
-import android.support.annotation.NonNull;
 import android.view.Gravity;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.PopupWindow;
 
+
+import androidx.annotation.NonNull;
 
 import com.trs.channellib.R;
 import com.trs.channellib.channel.channel.ChannelAdapter;
