@@ -1,10 +1,11 @@
 package com.jeremyfeinstein.slidingmenu.lib.app;
 
 import android.os.Bundle;
-import android.support.v4.app.FragmentActivity;
 import android.view.KeyEvent;
 import android.view.View;
 import android.view.ViewGroup.LayoutParams;
+
+import androidx.fragment.app.FragmentActivity;
 
 import com.jeremyfeinstein.slidingmenu.lib.SlidingMenu;
 
@@ -12,9 +13,6 @@ public class SlidingFragmentActivity extends FragmentActivity implements Sliding
 
 	private SlidingActivityHelper mHelper;
 
-	/* (non-Javadoc)
-	 * @see android.support.v4.app.FragmentActivity#onCreate(android.os.Bundle)
-	 */
 	@Override
 	public void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
@@ -42,9 +40,6 @@ public class SlidingFragmentActivity extends FragmentActivity implements Sliding
 		return mHelper.findViewById(id);
 	}
 
-	/* (non-Javadoc)
-	 * @see android.support.v4.app.FragmentActivity#onSaveInstanceState(android.os.Bundle)
-	 */
 	@Override
 	protected void onSaveInstanceState(Bundle outState) {
 		super.onSaveInstanceState(outState);
